(1) Fixed an issue in the CSU while using existing database as-is when only client tools are installed for copying in existing settings. Now checks if openPDC.exe.config exists before attempting to parse and falls back on openPDCManager.exe.config - check if it exists also before attempting parse, (2) Updated CSU such that database upgrades will suffix database name with major/minor version number of running PDQTracker
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS07]$/pdqtracker/Main;C96934

+++ Source/Tools/ConfigurationSetupUtility/Screens/SetupCompleteScreen.xaml.cs
@@ -375,9 +375,13 @@ namespace ConfigurationSetupUtility.Screens
         private void ValidateInternalDataPublisher()
         {
             string configFile = Directory.GetCurrentDirectory() + "\\PDQTracker.exe.config";
-            string configText = File.ReadAllText(configFile);
-            string replacedConfigText = configText.Replace("<datapublisher>", "<internaldatapublisher>").Replace("</datapublisher>", "</internaldatapublisher>");
-            File.WriteAllText(configFile, replacedConfigText);
+
+            if (File.Exists(configFile))
+            {
+                string configText = File.ReadAllText(configFile);
+                string replacedConfigText = configText.Replace("<datapublisher>", "<internaldatapublisher>").Replace("</datapublisher>", "</internaldatapublisher>");
+                File.WriteAllText(configFile, replacedConfigText);
+            }
         }
 
         private void ValidateGridSolutionsNamespaces()
